requirements.txt
fastapi
uvicorn
sqlite3
pydantic
python-jose
passlib